improved parsing of schema and tests
