+++ python/tests/unit/connectors/mcp/test_mcp.py
@@ -1,10 +1,9 @@
 # Copyright (c) Microsoft. All rights reserved.
-import os
 from typing import TYPE_CHECKING
 from unittest.mock import MagicMock, patch
 
 import pytest
-from mcp import ClientSession, StdioServerParameters
+from mcp import ClientSession, ListToolsResult, StdioServerParameters, Tool
 
 from semantic_kernel.connectors.mcp import (
     MCPSseServerConfig,
@@ -12,10 +11,9 @@ from semantic_kernel.connectors.mcp import (
     create_plugin_from_mcp_server,
 )
 from semantic_kernel.exceptions import KernelPluginInvalidConfigurationError
-from semantic_kernel.functions import KernelArguments
 
 if TYPE_CHECKING:
-    from semantic_kernel import Kernel
+    pass
 
 
 async def test_mcp_server_config_session_initialize():
@@ -104,34 +102,6 @@ async def test_mcp_stdio_server_config_failed_get_session():
                 pass
 
 
-async def test_from_mcp(kernel: "Kernel"):
-    mcp_server_path = os.path.join(os.path.dirname(__file__), "../../../assets/test_plugins", "TestMCPPlugin")
-    mcp_server_file = "mcp_server.py"
-    config = MCPStdioServerConfig(
-        command="uv",
-        args=["--directory", mcp_server_path, "run", mcp_server_file],
-    )
-
-    plugin = await create_plugin_from_mcp_server(
-        plugin_name="TestMCPPlugin",
-        description="Test MCP Plugin",
-        server_config=config,
-    )
-
-    assert plugin is not None
-    assert plugin.name == "TestMCPPlugin"
-    assert plugin.functions.get("get_name") is not None
-    assert plugin.functions["get_name"].parameters[0].name == "name"
-    assert plugin.functions["get_name"].parameters[0].type_ == "string"
-    assert plugin.functions["get_name"].parameters[0].is_required
-    assert plugin.functions.get("set_name") is not None
-
-    kernel.add_plugin(plugin)
-
-    result = await plugin.functions["get_name"].invoke(kernel, arguments=KernelArguments(name="test"))
-    assert "test: Test" in result.value
-
-
 @patch("semantic_kernel.connectors.mcp.stdio_client")
 @patch("semantic_kernel.connectors.mcp.ClientSession")
 async def test_with_kwargs_stdio(mock_session, mock_client):
@@ -145,7 +115,21 @@ async def test_with_kwargs_stdio(mock_session, mock_client):
 
     # Make the mock_stdio_client return an AsyncMock for the context manager
     mock_client.return_value = mock_generator
-    await create_plugin_from_mcp_server(
+    mock_session.return_value.__aenter__.return_value.list_tools.return_value = ListToolsResult(
+        tools=[
+            Tool(
+                name="get_name",
+                description="Get Name",
+                inputSchema={
+                    "properties": {
+                        "name": {"type": "string"},
+                    },
+                    "required": ["name"],
+                },
+            )
+        ]
+    )
+    plugin = await create_plugin_from_mcp_server(
         plugin_name="TestMCPPlugin",
         description="Test MCP Plugin",
         command="uv",
@@ -154,6 +138,12 @@ async def test_with_kwargs_stdio(mock_session, mock_client):
     mock_client.assert_called_once_with(
         server=StdioServerParameters(command="uv", args=["--directory", "path", "run", "file.py"])
     )
+    assert plugin is not None
+    assert plugin.name == "TestMCPPlugin"
+    assert plugin.description == "Test MCP Plugin"
+    assert plugin.functions.get("get_name") is not None
+    assert plugin.functions["get_name"].parameters[0].name == "name"
+    assert plugin.functions["get_name"].parameters[0].is_required
 
 
 @patch("semantic_kernel.connectors.mcp.sse_client")
@@ -169,9 +159,29 @@ async def test_with_kwargs_sse(mock_session, mock_client):
 
     # Make the mock_stdio_client return an AsyncMock for the context manager
     mock_client.return_value = mock_generator
-    await create_plugin_from_mcp_server(
+    mock_session.return_value.__aenter__.return_value.list_tools.return_value = ListToolsResult(
+        tools=[
+            Tool(
+                name="get_name",
+                description="Get Name",
+                inputSchema={
+                    "properties": {
+                        "name": {"type": "string"},
+                    },
+                    "required": ["name"],
+                },
+            )
+        ]
+    )
+    plugin = await create_plugin_from_mcp_server(
         plugin_name="TestMCPPlugin",
         description="Test MCP Plugin",
         url="http://localhost:8080/sse",
     )
     mock_client.assert_called_once_with(url="http://localhost:8080/sse")
+    assert plugin is not None
+    assert plugin.name == "TestMCPPlugin"
+    assert plugin.description == "Test MCP Plugin"
+    assert plugin.functions.get("get_name") is not None
+    assert plugin.functions["get_name"].parameters[0].name == "name"
+    assert plugin.functions["get_name"].parameters[0].is_required
